DONE with the cities

+++ Randemic_cod/RandomGen/procenteOras.h
@@ -1,5 +1,5 @@
-extern int counterMorti ;
-extern int counterVii;
+extern float counterMorti ;
+extern float counterVii;
 extern float counterInfectati;
-extern int counterVindecati;
+extern float counterVindecati;
